move stats up a class for nom
