+++ nmdc_ms_metadata_gen/lcms_nom_metadata_generator.py
@@ -1,7 +1,3 @@
-from typing import List
-
-import pandas as pd
-
 from nmdc_ms_metadata_gen.nom_metadata_generator import NOMMetadataGenerator
 
 
@@ -82,11 +78,6 @@ class LCMSNOMMetadataGenerator(NOMMetadataGenerator):
     )
     # on the mass spec records, you will need to add has_chromatograohy_configuration - created in parent metadata gen class
 
-    # QC thresholds
-    peak_count_threshold: int = 0
-    peak_assignment_count_threshold: int = 0
-    peak_assignment_rate_threshold: float = 0.0
-
     def __init__(
         self,
         metadata_file: str,
@@ -111,116 +102,6 @@ class LCMSNOMMetadataGenerator(NOMMetadataGenerator):
             workflow_version_git_url="https://raw.githubusercontent.com/microbiomedata/enviroMS/master/.bumpversion.cfg"
         )
         self.minting_config_creds = minting_config_creds
-
-    def generate_stats(self, processed_data: pd.DataFrame) -> List[int]:
-        """
-        Generate QC Stats from processed data.
-
-        Parameters
-        ----------
-        processed_data : pd.DataFrame
-            DataFrame containing the processed data.
-
-        Returns
-        -------
-        List[int]
-            List of QC stats generated from the processed data.
-
-        Notes
-        -----
-        This method reads in the processed data file and generates QC stats.
-
-        """
-
-        # Calculate peak_count
-        peak_count = processed_data["Index"].nunique()
-
-        # Calculate peak_assignment_count
-        peak_assignments = processed_data[["Index", "Molecular Formula"]].dropna()
-        peak_assignment_count = peak_assignments["Index"].nunique()
-
-        return peak_count, peak_assignment_count
-
-    def _get_wf_stats(self, processed_data: pd.DataFrame) -> dict:
-        """Return workflow statistics for di NOM data as a dict."""
-        peak_count, peak_assignment_count = self.generate_stats(
-            processed_data=processed_data
-        )
-        return {
-            "peak_count": peak_count,
-            "peak_assignment_count": peak_assignment_count,
-        }
-
-    def _resolve_qc_from_stats(self, qc_status, qc_comment, wf_stats: dict):
-        """Determine qc_status and qc_comment from di NOM stats and optional CSV input.
-
-        Threshold checks are always run. Resolution follows these rules:
-
-        1. Stats fail AND CSV says "fail": returns "fail" with both the CSV comment and
-           the stat failure message concatenated together.
-        2. Stats fail AND CSV says "pass" or no CSV input: stats prevail, returns "fail"
-           with the stat failure message only.
-        3. Stats pass AND CSV says "fail": CSV override is accepted unconditionally,
-           returns "fail" with the CSV comment only.
-        4. Stats pass AND no CSV "fail": returns "pass" with the CSV comment if provided,
-           otherwise the default pass message.
-
-        Parameters
-        ----------
-        qc_status : str or None
-            QC status value from the CSV, or None if not provided.
-        qc_comment : str or None
-            QC comment value from the CSV, or None if not provided.
-        wf_stats : dict
-            Dictionary of workflow statistics (peak_count, peak_assignment_count).
-
-        Returns
-        -------
-        tuple
-            A tuple of (qc_status, qc_comment) resolved according to the rules above.
-        """
-        # Always compute stat failures
-        failed = []
-        if wf_stats.get("peak_count", 0) < self.peak_count_threshold:
-            failed.append(
-                f"peak_count ({wf_stats.get('peak_count', 0)} < {self.peak_count_threshold})"
-            )
-        if (
-            wf_stats.get("peak_assignment_count", 0)
-            < self.peak_assignment_count_threshold
-        ):
-            failed.append(
-                f"peak_assignment_count ({wf_stats.get('peak_assignment_count', 0)} < {self.peak_assignment_count_threshold})"
-            )
-        peak_assignment_rate = (
-            wf_stats.get("peak_assignment_count", 0) / wf_stats.get("peak_count", 0)
-            if wf_stats.get("peak_count", 0) > 0
-            else 0
-        )
-        if peak_assignment_rate < self.peak_assignment_rate_threshold:
-            failed.append(
-                f"peak_assignment_rate ({peak_assignment_rate} < {self.peak_assignment_rate_threshold})"
-            )
-
-        stat_comment = f"QC failed on: {', '.join(failed)}." if failed else None
-
-        if failed and qc_status == "fail":
-            # Both stats and CSV indicate failure — concatenate comments
-            combined_comment = "; ".join(filter(None, [qc_comment, stat_comment]))
-            return "fail", combined_comment
-        elif failed:
-            # Stats fail, CSV says "pass" or nothing — stats prevail
-            return "fail", stat_comment
-        elif qc_status == "fail":
-            # Stats pass, but CSV explicitly forces a fail — accept it
-            return qc_status, qc_comment
-        else:
-            # Stats pass and no CSV override to fail
-            return "pass", (
-                qc_comment
-                if qc_comment is not None
-                else "QC passed all computed peak count thresholds."
-            )
 
     def rerun(self):
         return super().rerun()
